[NoteUtils] Add DataBase64Encoded and ContentID fields to Attachment
KPIMTextEdit returns already base64-encoded images, so if I want to attach them,
I have to decode it, just for KMime to encode them again. Instead, just
make it possible to explicitly tell KMime that attachment is already
base64-encoded.

It is also useful to be able to set some unique identifier to an
Attachment, just to be able to refer to it from the note (`contentID`).
Right now we only have `label` field, which is not unique and is
supposed to be user-facing.